+++ .github/workflows/nightly-tests.yml
@@ -166,7 +166,7 @@ jobs:
         uses: actions/checkout@v4
         with:
           repository: powdr-labs/openvm-reth-benchmark
-          ref: main
+          ref: 29dc4c3581acbc3818543bef360e17a5938ff19b
           path: openvm-reth-benchmark
 
       - name: Patch openvm-reth-benchmark to use local powdr
