+++ autoprecompiles/Cargo.toml
@@ -16,6 +16,9 @@ log = "0.4.18"
 num-traits = "0.2.15"
 serde = "1.0.218"
 serde_cbor = "0.11.2"
+tracing = "0.1.41"
+serde_json = "1.0.140"
+rayon = "1.10.0"
 
 [package.metadata.cargo-udeps.ignore]
 development = ["env_logger"]
