+++ openvm/Cargo.toml
@@ -61,7 +61,6 @@ log = "0.4.17"
 struct-reflection = { git = "https://github.com/gzanitti/struct-reflection-rs.git" }
 
 toml = "0.8.14"
-serde_json = "1.0.140"
 
 [dev-dependencies]
 test-log = { version = "0.2.17", features = ["trace"] }
