+++ openvm/src/customize_exe.rs
@@ -1,32 +1,33 @@
 use std::collections::{BTreeSet, HashMap};
 
-use std::io::BufWriter;
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
+use std::path::Path;
+use std::sync::Arc;
 
 use crate::extraction_utils::{get_air_metrics, OriginalAirs, OriginalVmConfig};
 use crate::instruction_formatter::openvm_instruction_formatter;
 use crate::opcode::{branch_opcodes_bigint_set, branch_opcodes_set};
 use crate::powdr_extension::chip::PowdrAir;
-use crate::utils::{fractional_knapsack, KnapsackItem, UnsupportedOpenVmReferenceError};
-use crate::IntoOpenVm;
+use crate::utils::UnsupportedOpenVmReferenceError;
 use crate::OpenVmField;
 use crate::OriginalCompiledProgram;
 use crate::{CompiledProgram, SpecializedConfig};
+use crate::{IntoOpenVm, PrecompileImplementation};
 use itertools::Itertools;
 use openvm_instructions::instruction::Instruction;
 use openvm_instructions::program::Program as OpenVmProgram;
 use openvm_instructions::VmOpcode;
-use openvm_stark_backend::p3_maybe_rayon::prelude::*;
 use openvm_stark_backend::{
     interaction::SymbolicInteraction,
     p3_field::{FieldAlgebra, PrimeField32},
 };
-use powdr_autoprecompiles::basic_blocks::{collect_basic_blocks, Program};
+use powdr_autoprecompiles::blocks::{collect_basic_blocks, Program};
+use powdr_autoprecompiles::blocks::{generate_apcs_with_pgo, Candidate, KnapsackItem, PgoConfig};
+use powdr_autoprecompiles::constraint_optimizer::IsBusStateful;
 use powdr_autoprecompiles::expression::try_convert;
-use powdr_autoprecompiles::{Apc, DegreeBound};
+use powdr_autoprecompiles::{Apc, PowdrConfig};
 use powdr_autoprecompiles::{BasicBlock, VmConfig};
 use powdr_autoprecompiles::{SymbolicBusInteraction, SymbolicInstructionStatement};
+use powdr_constraint_solver::constraint_system::BusInteractionHandler;
 use powdr_number::{BabyBearField, FieldElement};
 use serde::{Deserialize, Serialize};
 
@@ -36,11 +37,7 @@ use crate::{
     utils::symbolic_to_algebraic,
 };
 
-pub const OPENVM_DEGREE_BOUND: usize = 5;
-
 pub const POWDR_OPCODE: usize = 0x10ff;
-
-use crate::{PgoConfig, PowdrConfig};
 
 #[derive(Debug)]
 pub enum Error {
@@ -53,37 +50,6 @@ impl From<powdr_autoprecompiles::constraint_optimizer::Error> for Error {
     }
 }
 
-fn generate_apcs_with_pgo(
-    blocks: Vec<BasicBlock<BabyBearField>>,
-    config: &PowdrConfig,
-    original_config: &OriginalVmConfig,
-    pgo_config: PgoConfig,
-    vm_config: VmConfig<OriginalAirs<BabyBearField>, OpenVmBusInteractionHandler<BabyBearField>>,
-) -> Vec<Apc<BabyBearField>> {
-    // sort basic blocks by:
-    // 1. if PgoConfig::Cell, cost = frequency * cells_saved_per_row
-    // 2. if PgoConfig::Instruction, cost = frequency * number_of_instructions
-    // 3. if PgoConfig::None, cost = number_of_instructions
-    let res = match pgo_config {
-        PgoConfig::Cell(pgo_program_idx_count, max_total_columns) => create_apcs_with_cell_pgo(
-            blocks,
-            pgo_program_idx_count,
-            max_total_columns,
-            config,
-            original_config,
-            vm_config,
-        ),
-        PgoConfig::Instruction(pgo_program_idx_count) => {
-            create_apcs_with_instruction_pgo(blocks, pgo_program_idx_count, config, vm_config)
-        }
-        PgoConfig::None => create_apcs_with_no_pgo(blocks, config, vm_config),
-    };
-
-    assert!(res.len() <= config.autoprecompiles as usize);
-
-    res
-}
-
 pub fn customize(
     OriginalCompiledProgram {
         mut exe,
@@ -91,6 +57,7 @@ pub fn customize(
     }: OriginalCompiledProgram,
     labels: &BTreeSet<u32>,
     config: PowdrConfig,
+    implementation: PrecompileImplementation,
     pgo_config: PgoConfig,
 ) -> CompiledProgram {
     let original_config = OriginalVmConfig::new(sdk_vm_config.clone());
@@ -131,6 +98,18 @@ pub fn customize(
         bus_map: bus_map.clone(),
     };
 
+    let max_total_apc_columns: Option<usize> = match pgo_config {
+        PgoConfig::Cell(_, max_total_columns) => max_total_columns.map(|max_total_columns| {
+            let total_non_apc_columns = original_config
+                .chip_inventory_air_metrics()
+                .values()
+                .map(|m| m.total_width())
+                .sum::<usize>();
+            max_total_columns - total_non_apc_columns
+        }),
+        PgoConfig::Instruction(_) | PgoConfig::None => None,
+    };
+
     let blocks = collect_basic_blocks(&program, &labels, &opcodes_allowlist, &branch_opcodes_set());
     tracing::info!(
         "Got {} basic blocks from `collect_basic_blocks`",
@@ -166,7 +145,13 @@ pub fn customize(
         })
         .collect::<Vec<_>>();
 
-    let apcs = generate_apcs_with_pgo(blocks, &config, &original_config, pgo_config, vm_config);
+    let apcs = generate_apcs_with_pgo::<OpenVmApcCandidate<_>, _, _, _>(
+        blocks,
+        &config,
+        max_total_apc_columns,
+        pgo_config,
+        vm_config,
+    );
 
     let program = &mut exe.program.instructions_and_debug_infos;
 
@@ -248,7 +233,7 @@ pub fn customize(
 
     CompiledProgram {
         exe,
-        vm_config: SpecializedConfig::new(original_config, extensions, config.implementation),
+        vm_config: SpecializedConfig::new(original_config, extensions, implementation),
     }
 }
 
@@ -282,47 +267,6 @@ fn add_extra_targets<F: PrimeField32>(
     labels
 }
 
-// Only used for PgoConfig::Instruction and PgoConfig::None,
-// because PgoConfig::Cell caches all APCs in sorting stage.
-fn create_apcs_for_all_blocks<P: FieldElement>(
-    blocks: Vec<BasicBlock<P>>,
-    powdr_config: &PowdrConfig,
-    vm_config: VmConfig<OriginalAirs<P>, OpenVmBusInteractionHandler<P>>,
-) -> Vec<Apc<P>> {
-    let n_acc = powdr_config.autoprecompiles as usize;
-    tracing::info!("Generating {n_acc} autoprecompiles in parallel");
-
-    blocks
-        .into_par_iter()
-        .skip(powdr_config.skip_autoprecompiles as usize)
-        .take(n_acc)
-        .enumerate()
-        .map(|(index, block)| {
-            tracing::debug!(
-                "Accelerating block of length {} and start idx {}",
-                block.statements.len(),
-                block.start_idx
-            );
-
-            tracing::debug!(
-                "Acc block: {}",
-                block.pretty_print(openvm_instruction_formatter)
-            );
-
-            let apc_opcode = POWDR_OPCODE + index;
-
-            powdr_autoprecompiles::build(
-                block,
-                vm_config.clone(),
-                powdr_config.degree_bound,
-                apc_opcode as u32,
-                powdr_config.apc_candidates_dir_path.as_deref(),
-            )
-            .unwrap()
-        })
-        .collect()
-}
-
 pub fn openvm_bus_interaction_to_powdr<F: PrimeField32, P: FieldElement>(
     interaction: &SymbolicInteraction<F>,
     columns: &[Arc<String>],
@@ -340,51 +284,23 @@ pub fn openvm_bus_interaction_to_powdr<F: PrimeField32, P: FieldElement>(
 }
 
 #[derive(Serialize, Deserialize)]
-struct ApcCandidate<P> {
+struct OpenVmApcCandidate<P> {
     apc: Apc<P>,
     execution_frequency: usize,
     width_before: usize,
     width_after: usize,
 }
 
-#[derive(Serialize, Deserialize)]
-struct ApcCandidateJsonExport<P> {
-    // opcode
-    opcode: u32,
-    // execution_frequency
-    execution_frequency: usize,
-    // original instructions
-    original_block: BasicBlock<P>,
-    // total width before optimisation
-    total_width_before: usize,
-    // total width after optimisation
-    total_width_after: usize,
-    // path to the apc candidate file
-    apc_candidate_file: String,
-}
+impl<B: BusInteractionHandler<BabyBearField> + Clone + Sync + IsBusStateful<BabyBearField>>
+    Candidate<BabyBearField, OriginalAirs<BabyBearField>, B> for OpenVmApcCandidate<BabyBearField>
+{
+    type JsonExport = OpenVmApcCandidateJsonExport<BabyBearField>;
 
-impl ApcCandidate<BabyBearField> {
-    /// Try to create an autoprecompile candidate from a block.
-    pub fn try_create(
-        block: BasicBlock<BabyBearField>,
-        opcode: usize,
-        degree_bound: DegreeBound,
+    fn create(
+        apc: Apc<BabyBearField>,
         pgo_program_idx_count: &HashMap<u32, u32>,
-        apc_candidates_dir_path: &Option<PathBuf>,
-        vm_config: VmConfig<
-            OriginalAirs<BabyBearField>,
-            OpenVmBusInteractionHandler<BabyBearField>,
-        >,
-    ) -> Option<Self> {
-        let apc = powdr_autoprecompiles::build(
-            block,
-            vm_config.clone(),
-            degree_bound,
-            opcode as u32,
-            apc_candidates_dir_path.as_deref(),
-        )
-        .ok()?;
-
+        vm_config: VmConfig<OriginalAirs<BabyBearField>, B>,
+    ) -> Self {
         let apc_metrics = get_air_metrics(Arc::new(PowdrAir::new(apc.machine().clone())));
         let width_after = apc_metrics.widths.total();
 
@@ -406,22 +322,20 @@ impl ApcCandidate<BabyBearField> {
             .get(&(apc.block.start_idx as u32))
             .unwrap_or(&0) as usize;
 
-        let candidate = Self {
+        Self {
             apc,
             execution_frequency,
             width_before,
             width_after,
-        };
-
-        Some(candidate)
+        }
     }
 
     /// Return a JSON export of the APC candidate.
     fn to_json_export(
         &self,
         apc_candidates_dir_path: &Path,
-    ) -> ApcCandidateJsonExport<BabyBearField> {
-        ApcCandidateJsonExport {
+    ) -> OpenVmApcCandidateJsonExport<BabyBearField> {
+        OpenVmApcCandidateJsonExport {
             opcode: self.apc.opcode,
             execution_frequency: self.execution_frequency,
             original_block: self.apc.block.clone(),
@@ -433,16 +347,36 @@ impl ApcCandidate<BabyBearField> {
                 .to_string(),
         }
     }
+
+    fn into_apc(self) -> Apc<BabyBearField> {
+        self.apc
+    }
 }
 
-impl<P> ApcCandidate<P> {
+#[derive(Serialize, Deserialize)]
+struct OpenVmApcCandidateJsonExport<P> {
+    // opcode
+    opcode: u32,
+    // execution_frequency
+    execution_frequency: usize,
+    // original instructions
+    original_block: BasicBlock<P>,
+    // total width before optimisation
+    total_width_before: usize,
+    // total width after optimisation
+    total_width_after: usize,
+    // path to the apc candidate file
+    apc_candidate_file: String,
+}
+
+impl<P> OpenVmApcCandidate<P> {
     fn cells_saved_per_row(&self) -> usize {
         // The number of cells saved per row is the difference between the width before and after the APC.
         self.width_before - self.width_after
     }
 }
 
-impl<P> KnapsackItem for ApcCandidate<P> {
+impl<P> KnapsackItem for OpenVmApcCandidate<P> {
     fn cost(&self) -> usize {
         self.width_after
     }
@@ -461,153 +395,4 @@ impl<P> KnapsackItem for ApcCandidate<P> {
     fn tie_breaker(&self) -> usize {
         self.apc.opcode as usize
     }
-}
-
-// Note: This function can lead to OOM since it generates the apc for many blocks.
-fn create_apcs_with_cell_pgo(
-    mut blocks: Vec<BasicBlock<BabyBearField>>,
-    pgo_program_idx_count: HashMap<u32, u32>,
-    max_total_columns: Option<usize>,
-    config: &PowdrConfig,
-    original_config: &OriginalVmConfig,
-    vm_config: VmConfig<OriginalAirs<BabyBearField>, OpenVmBusInteractionHandler<BabyBearField>>,
-) -> Vec<Apc<BabyBearField>> {
-    // drop any block whose start index cannot be found in pc_idx_count,
-    // because a basic block might not be executed at all.
-    // Also only keep basic blocks with more than one original instruction.
-    blocks.retain(|b| {
-        pgo_program_idx_count.contains_key(&(b.start_idx as u32)) && b.statements.len() > 1
-    });
-
-    tracing::debug!(
-        "Retained {} basic blocks after filtering by pc_idx_count",
-        blocks.len()
-    );
-
-    // generate apc for all basic blocks and only cache the ones we eventually use
-    // calculate number of trace cells saved per row for each basic block to sort them by descending cost
-    let max_cache = (config.autoprecompiles + config.skip_autoprecompiles) as usize;
-    tracing::info!(
-        "Generating autoprecompiles for all ({}) basic blocks in parallel and caching costliest {}",
-        blocks.len(),
-        max_cache,
-    );
-
-    let max_total_apc_columns: Option<usize> = max_total_columns.map(|max_total_columns| {
-        let total_non_apc_columns = original_config
-            .chip_inventory_air_metrics()
-            .values()
-            .map(|m| m.total_width())
-            .sum::<usize>();
-        max_total_columns - total_non_apc_columns
-    });
-
-    let apc_candidates = Arc::new(Mutex::new(vec![]));
-
-    // map–reduce over blocks into a single BinaryHeap<ApcCandidate<P>> capped at max_cache
-    let res = fractional_knapsack(
-        blocks.into_par_iter().enumerate().filter_map(|(i, block)| {
-            ApcCandidate::try_create(
-                block,
-        POWDR_OPCODE + i,
-                config.degree_bound,
-                &pgo_program_idx_count,
-                &config.apc_candidates_dir_path,
-                vm_config.clone(),
-            ).inspect(|candidate| {
-                if let Some(apc_candidates_dir_path) = &config.apc_candidates_dir_path {
-                    let json_export = candidate.to_json_export(apc_candidates_dir_path);
-                    apc_candidates.lock().unwrap().push(json_export);
-                }
-            })
-        }),
-        max_cache,
-        max_total_apc_columns,
-    )
-    .skip(config.skip_autoprecompiles as usize)
-    .map(|c| {
-        tracing::debug!(
-            "Basic block start_idx: {}, cost adjusted value: {}, frequency: {}, cells_saved_per_row: {}",
-            c.apc.block.start_idx,
-            c.value() / c.cost(),
-            c.execution_frequency,
-            c.cells_saved_per_row(),
-        );
-
-        c.apc
-    })
-    .collect();
-
-    // Write the APC candidates JSON to disk if the directory is specified.
-    if let Some(apc_candidates_dir_path) = &config.apc_candidates_dir_path {
-        let apc_candidates_json_file = apc_candidates.lock().unwrap();
-        let json_path = apc_candidates_dir_path.join("apc_candidates.json");
-        let file = std::fs::File::create(&json_path)
-            .expect("Failed to create file for APC candidates JSON");
-        serde_json::to_writer(BufWriter::new(file), &*apc_candidates_json_file)
-            .expect("Failed to write APC candidates JSON to file");
-    }
-
-    res
-}
-
-fn create_apcs_with_instruction_pgo<P: FieldElement>(
-    mut blocks: Vec<BasicBlock<P>>,
-    pgo_program_idx_count: HashMap<u32, u32>,
-    config: &PowdrConfig,
-    vm_config: VmConfig<OriginalAirs<P>, OpenVmBusInteractionHandler<P>>,
-) -> Vec<Apc<P>> {
-    // drop any block whose start index cannot be found in pc_idx_count,
-    // because a basic block might not be executed at all.
-    // Also only keep basic blocks with more than one original instruction.
-    blocks.retain(|b| {
-        pgo_program_idx_count.contains_key(&(b.start_idx as u32)) && b.statements.len() > 1
-    });
-
-    tracing::debug!(
-        "Retained {} basic blocks after filtering by pc_idx_count",
-        blocks.len()
-    );
-
-    // cost = cells_saved_per_row
-    blocks.sort_by(|a, b| {
-        let a_cnt = pgo_program_idx_count[&(a.start_idx as u32)];
-        let b_cnt = pgo_program_idx_count[&(b.start_idx as u32)];
-        (b_cnt * (b.statements.len() as u32)).cmp(&(a_cnt * (a.statements.len() as u32)))
-    });
-
-    // Debug print blocks by descending cost
-    for block in &blocks {
-        let start_idx = block.start_idx;
-        let frequency = pgo_program_idx_count[&(start_idx as u32)];
-        let number_of_instructions = block.statements.len();
-        let value = frequency * number_of_instructions as u32;
-
-        tracing::debug!(
-            "Basic block start_idx: {start_idx}, value: {value}, frequency: {frequency}, number_of_instructions: {number_of_instructions}",
-        );
-    }
-
-    create_apcs_for_all_blocks(blocks, config, vm_config)
-}
-
-fn create_apcs_with_no_pgo<P: IntoOpenVm>(
-    mut blocks: Vec<BasicBlock<P>>,
-    config: &PowdrConfig,
-    vm_config: VmConfig<OriginalAirs<P>, OpenVmBusInteractionHandler<P>>,
-) -> Vec<Apc<P>> {
-    // cost = number_of_original_instructions
-    blocks.sort_by(|a, b| b.statements.len().cmp(&a.statements.len()));
-
-    // Debug print blocks by descending cost
-    for block in &blocks {
-        let start_idx = block.start_idx;
-        tracing::debug!(
-            "Basic block start_idx: {}, number_of_instructions: {}",
-            start_idx,
-            block.statements.len(),
-        );
-    }
-
-    create_apcs_for_all_blocks(blocks, config, vm_config)
 }
